move style loading into init(). Because we have to wait until loaded the style from the settings
